backend/pyproject.toml
[project]
name = "remote-radar"
version = "0.1.0"
description = ""
authors = [
    {name = "Ista2000",email = "mishra.istasis@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "beautifulsoup4 (>=4.13.3,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "langchain-groq (>=0.3.2,<0.4.0)",
    "fastapi (==0.115.9)",
    "langchain-community (>=0.3.21,<0.4.0)",
    "chromadb (>=1.0.4,<2.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
pre-commit = "^4.2.0"
black = "^25.1.0"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
